Prepare for multiple subtasks
