+++ pyversions/common.py
@@ -1,0 +1,7 @@
+import logging
+
+log = logging.getLogger('python-versions')
+log.setLevel(logging.DEBUG)
+log.addHandler(logging.NullHandler())
+
+BUG_URL = 'https://github.com/fedora-python/task-python-versions/issues'
