+++ test/functional/test_two_three.py
@@ -3,7 +3,7 @@ import os
 
 import pytest
 
-from pyversions import two_three_check
+from pyversions.two_three import check_two_three
 
 
 def pkg(filename):
@@ -17,26 +17,26 @@ def gpkg(pkgglob):
 @pytest.mark.parametrize('pkgglob', ('pyserial*', 'python-peak-rules*',
                                      'python2-geoip2*'))
 def test_package_depends_on_2_only(pkgglob):
-    name, versions = two_three_check(gpkg(pkgglob))
+    name, versions = check_two_three(gpkg(pkgglob))
     assert 2 in versions
     assert 3 not in versions
 
 
 @pytest.mark.parametrize('pkgglob', ('python3-pyserial*',))
 def test_package_depends_on_3_only(pkgglob):
-    name, versions = two_three_check(gpkg(pkgglob))
+    name, versions = check_two_three(gpkg(pkgglob))
     assert 2 not in versions
     assert 3 in versions
 
 
 @pytest.mark.parametrize('pkgglob', ('tracer*',))
 def test_package_depends_on_2_and_3(pkgglob):
-    name, versions = two_three_check(gpkg(pkgglob))
+    name, versions = check_two_three(gpkg(pkgglob))
     assert 2 in versions
     assert 3 in versions
 
 
 @pytest.mark.parametrize('pkgglob', ('libgccjit-devel*',))
 def test_package_depends_on_no_python(pkgglob):
-    name, versions = two_three_check(gpkg(pkgglob))
+    name, versions = check_two_three(gpkg(pkgglob))
     assert not versions
